==> code/notebooks/01_BAIS3250_project_cleaningAPIdata.ipynb <==
# BAIS:3250 - Final Project
### Cleaning API Data

**Author(s):** Natalie Brown, Max Kaiser

**Date Modified:** 11-13-2024 (*date created:* 11-13-2024)


**Description:** Clean country data that was scraped from RESTcountries API

---

### Import Libaries
* **pandas:** for data frames and data cleaning functions

import pandas as pd

---
### Load Data
* **api_country_data_raw.csv**

# example load code
country_df=pd.read_csv('00_api_country_data_raw.csv',sep=',',encoding='utf-8',header=0)

# display head to ensure properly loaded
country_df.head()

---
### Perform intial discovery
* display datatypes
* display nulls
* display data frame size

# data frame size
shape=country_df.shape
print(f'Country Data Shape\n------------------\nColumns: {shape[1]}\nRows: {shape[0]}')

# display data types
print(f'Country Data Types\n------------------\n{country_df.dtypes}')

# determine nulls counts and percentages
null_counts=country_df.isna().sum()
null_percentages=round((country_df.isna().sum()/shape[0])*100,2)

# data frame for results
nulls_df=pd.DataFrame({
    'Nulls':null_counts,
    'Nulls (%)':null_percentages
})

# display
print("Country Data Nulls")
nulls_df

---
### Perform Inital Cleaning
- [x] remove columns with over 50% nulls
- [x] fix data types
    * displaying values first to get formatting
    * features to be fixed (all need to be integer):
        - [x] **covid_cases**
        - [x] **covid_deaths**
        - [x] **size** renamed to **area_km2**
        - [x] **population**
- [x] fill continent for the two countries missing

#### Remove Columns

# remove columns
country_df=country_df.drop(columns=['president_name','president_gender','president_appointment_start_date','president_appointment_end_date','capital'])

#### Fix data types

# function to display current data format
def data_format(col):
    print(f'Data Format for {col}\n-----------------\n{country_df[col].unique}')

**covid_cases**

data_format('covid_cases')

# remove commas and fix to integer
country_df['covid_cases_edited']=country_df['covid_cases'].str.replace(',', '').astype(int)

# display changes to confirm
country_df[['covid_cases_edited','covid_cases']]

# remove the original and rename the old
country_df=country_df.drop(columns='covid_cases')

# rename edited
country_df=country_df.rename(columns={'covid_cases_edited':'covid_cases'})

# review changes
country_df['covid_cases']

***covid_cases** complete*

**covid_deaths**

data_format('covid_deaths')

# remove commas and fix to integer
country_df['covid_deaths_edited']=country_df['covid_deaths'].str.replace(',', '').astype(int)

# display changes to confirm
country_df[['covid_deaths_edited','covid_deaths']]

# remove the original and rename the old
country_df=country_df.drop(columns='covid_deaths')

# rename edited
country_df=country_df.rename(columns={'covid_deaths_edited':'covid_deaths'})

# review changes
country_df['covid_deaths']

***covid_deaths** complete*

**size**

data_format('size')

# remove commas and measure
country_df['size_edited']=country_df['size'].str.replace(',', '').str.replace(' km²', '').astype(int)

# display changes to confirm
country_df[['size_edited','size']]

# remove the original and rename the old
country_df=country_df.drop(columns='size')

# rename edited, add a note about the measure
country_df=country_df.rename(columns={'size_edited':'area_km2'})

# review changes
country_df['area_km2']

***size** completed, renamed to **area_km2***

**population**

data_format('population')

# remove commas
country_df['population_edited']=country_df['population'].str.replace(',', '').astype(int)

# display changes to confirm
country_df[['population_edited','population']]

# remove the original and rename the old
country_df=country_df.drop(columns='population')

# rename edited
country_df=country_df.rename(columns={'population_edited':'population'})

# review changes
country_df['population']

***population** completed*

**fill missing continents**

# first find the occurences
null_continent_rows=country_df.loc[country_df['continent'].isna()]

null_continent_rows

country_df.loc[country_df['country']=='Trinidad And Tobago', 'continent']=country_df.loc[country_df['country']=='Trinidad And Tobago', 'continent'].fillna('North America')
country_df.loc[country_df['country']=='Tunisia', 'continent']=country_df.loc[country_df['country']=='Tunisia', 'continent'].fillna('Africa')

# review changes
country_df[country_df['country']=='Trinidad And Tobago']

country_df[country_df['country']=='Tunisia']

# display nulls again
null_counts=country_df.isna().sum()
null_percentages=round((country_df.isna().sum()/shape[0])*100,2)

# data frame for results
nulls_df=pd.DataFrame({
    'Nulls':null_counts,
    'Nulls (%)':null_percentages
})

# display
print("Country Data Nulls (cleaned)")
nulls_df

# display head one last time
country_df.head()

---
### Save Data
* **api_country_data_clean.csv**

# save to csv
country_df.to_csv('01_api_country_data_clean.csv',sep=',',encoding='utf-8',header=True,index=False)
