==> code/notebooks/02_BAIS3250_project_cleaningWorldData.ipynb <==
# BAIS:3250 - Final Project
### Cleaning World Data

**Author(s):** Natalie Brown, Max Kaiser

**Date Modified:** 11-14-2024 (*date created:* 11-13-2024)


**Description:** Clean world data from csv file that was downloaded from Kaggle

---

### Import Libaries
* **pandas:** for data frames and data cleaning functions
* **numpy** for null functions

import pandas as pd
import numpy as np

---
### Load Data
* **world-data-2023.csv**

# load file
world_df=pd.read_csv('02_world_data_raw.csv',sep=',',encoding='utf-8',header=0)

# display header
world_df.head()

---
### Perform Initial Discovery
- [x] data frame size
- [x] columns
- [x] nulls
- [x] data types

# data frame size
shape=world_df.shape
print(f'World Data Shape\n------------------\nColumns: {shape[1]}\nRows: {shape[0]}')

# columns
print(f'World Data Columns\n-------------------\n{world_df.columns}')

# data types
print(f'World Data Types\n------------------------\n{world_df.dtypes}')

# nulls
# display nulls again
null_counts=world_df.isna().sum()
null_percentages=round((world_df.isna().sum()/shape[0])*100,2)

# data frame for results
nulls_df=pd.DataFrame({
    'Nulls':null_counts,
    'Nulls (%)':null_percentages
})

# display
print("World Data Nulls")
nulls_df

---
### Perform Inital Cleaning
- [x] remove unnecessary columns
- [x] rename columns (snake case)
- [x] fix data types

*null imputation will be completed in a different notebook*

---
**Remove Unecessary Columns**
- [x] Abbreviation - do not need abbreviation, have country name
- [x] Armed Forces size - unecessary
- [x] Calling Code - do not need calling code, have country name
- [x] Largest city - unecessary
- [x] Density\n(P/Km2) - unecessary
- [x] Total tax rate - tax rates are not a good comparative measure, will use tax revenue %
- [x] Minimum wage - the cost of living varies by country so this is a bad measure
- [x] Gasoline Price - not a good measure as cost of living varies
- [x] Population: Labor force participation (%) - using unemployment rate for this measure instead
- [x] Out of pocket health expenditure - bad measure as many variables impact this across different countries
- [x] Co2-Emissions - unecessary measure
- [x] Forested Area (%) - bad measure
- [x] Urban_population - unecessary measure
- [x] CPI Change (%) - unecessary measure
- [x] Maternal mortality ratio - unecessary measure
- [x] Capital/Major City - unecessary measure

# columns to remove: calling code, capital / major city & land area (other data set has this and does not have nulls), density, abbreviation, armed forces size
world_df=world_df.drop(columns=['Abbreviation','Density\n(P/Km2)','Armed Forces size',
                                'Calling Code','Largest city','Total tax rate','Minimum wage',
                               'Gasoline Price','Population: Labor force participation (%)','Out of pocket health expenditure',
                               'Co2-Emissions','Forested Area (%)','Urban_population','CPI Change (%)',
                               'Maternal mortality ratio','Capital/Major City'])

display(world_df.head())

---
**Rename Columns**
- [x] Country : country
- [x] Agricultural Land( %) : ag_land_pct
- [x] Land Area(Km2) : land_area_km2
- [x] Birth Rate : birth_rt
- [x] CPI : consumer_price_index
- [x] Currency-Code : currency_code
- [x] Fertility Rate : fertility_rt
- [x] GDP : gross_domestic_product
- [x] Gross primary education enrollment (%) : primary_education_enrollment_pct
- [x] Gross tertiary education enrollment (%) : seconary_education_enrollment_pct
- [x] Infant mortality : infant_mortality_rt
- [x] Life expectancy : life_expectancy
- [x] Official language : official_language
- [x] Physicians per thousand : physicians_per_thousand
- [x] Population : population
- [x] Tax revenue (%) : tax_revenue_pct
- [x] Unemployment rate : unemployment_rt
- [x] Latitude : lat
- [x] Longitude : long

# rename
world_df=world_df.rename(columns={
    'Country':'country',
    'Agricultural Land( %)':'ag_land_pct',
    'Land Area(Km2)':'land_area_km2',
    'Birth Rate':'birth_rt',
    'CPI':'consumer_price_index',
    'Currency-Code':'currency_code',
    'Fertility Rate':'fertility_rt',
    'GDP':'gross_domestic_product',
    'Gross primary education enrollment (%)':'primary_education_enrollment_pct',
    'Gross tertiary education enrollment (%)':'secondary_education_enrollment_pct',
    'Infant mortality':'infant_mortality_rt',
    'Life expectancy':'life_expectancy',
    'Official language':'official_language',
    'Physicians per thousand':'physicians_per_thousand',
    'Population':'population',
    'Tax revenue (%)':'tax_revenue_pct',
    'Unemployment rate':'unemployment_rt',
    'Latitude':'lat',
    'Longitude':'long'
})

# review changes
world_df.head()

---
**Fix Data Types**
- [x] **ag_land_pct:** object - float
- [x] **land_area_km2:** object - float
- [x] **consumer_price_index:** object - float
- [x] **gross_domestic_product:** object - float
    - renamed to **gross_domestic_product_USD(b)** to inform the numbers are in billions, and measured in USD
- [x] **primary_education_enrollment_pct:** object - float
- [x] **seconary_education_enrollment_pct:** object - float
- [x] **population:** object - int
- [x] **tax_revenue_pct:** object - float
- [x] **unemployment_rt:** object - float

# display data types again, now that we have removed columns
print(f'World Data Types\n------------------------\n{world_df.dtypes}')

# formula to display current format, and nulls if any
def format(col):
    print(f'Nulls in {col}\n-----------------------\n{world_df[col].isna().sum()}\n\nCurrent Format of {col}\n--------------------\n{world_df[col].unique()}')

# formula to display data types changes
def type_change(col):
    edited=col+'_edited'
    print(f'Data Type Before: {world_df[col].dtype}\nData Type After: {world_df[edited].dtype}')
    display(world_df[[edited,col]].head())

def name_change(col):
    display(world_df[col].head())

**ag_land_pct**

object - float

format('ag_land_pct')

# remove percentage and convert to float
world_df['ag_land_pct_edited']=world_df['ag_land_pct'].str.replace('%','').astype(float)

type_change('ag_land_pct')

# rename and drop
world_df=world_df.drop(columns='ag_land_pct').rename(columns={'ag_land_pct_edited':'ag_land_pct'})

# review drop and rename
name_change('ag_land_pct')

***ag_land_pct** completed*

**land_area_km2**

object - float

# display format
format('land_area_km2')

# remove comma and convert to float
world_df['land_area_km2_edited']=world_df['land_area_km2'].str.replace(',','').astype(float)

# display changes
type_change('land_area_km2')

# rename and drop
world_df=world_df.drop(columns='land_area_km2').rename(columns={'land_area_km2_edited':'land_area_km2'})

# review drop and rename
name_change('land_area_km2')

 ***land_area_km2** completed*

**consumer_price_index**

object - float

# displau format and nulls
format('consumer_price_index')

# convert to int, ignore erros
world_df['consumer_price_index_edited']=world_df['consumer_price_index'].str.replace(',','').astype(float)

# review change
type_change('consumer_price_index')

# rename and drop
world_df=world_df.drop(columns='consumer_price_index').rename(columns={'consumer_price_index_edited':'consumer_price_index'})

# review drop and rename
name_change('consumer_price_index')

***consumer_price_index** completed*

**gross_domestic_product**

object - float

# review format and nulls
format('gross_domestic_product')

# review commas and $, convert to float and then convert to billions
world_df['gross_domestic_product_edited']=round((world_df['gross_domestic_product'].str.replace(',','').str.replace('$','').astype(float)/1000000000),4)

# review change
type_change('gross_domestic_product')

# drop, rename, make a note about the measures, 
world_df=world_df.drop(columns='gross_domestic_product').rename(columns={'gross_domestic_product_edited':'gross_domestic_product_USD(b)'})

# review change
name_change('gross_domestic_product_USD(b)')

***gross_domestic_product_USD(b)** completed*

**primary_education_enrollment_pct**

object - float

# display format
format('primary_education_enrollment_pct')

# remove %
world_df['primary_education_enrollment_pct_edited']=world_df['primary_education_enrollment_pct'].str.replace('%','').astype(float)

# review changes
type_change('primary_education_enrollment_pct')

# change names
world_df=world_df.drop(columns='primary_education_enrollment_pct').rename(columns={
    'primary_education_enrollment_pct_edited':'primary_education_enrollment_pct'})

# review changes
name_change('primary_education_enrollment_pct')

***primary_education_enrollment_pct** completed*

**secondary_education_enrollment_pct**

object - float

# display format
format('secondary_education_enrollment_pct')

# remove %
world_df['secondary_education_enrollment_pct_edited']=world_df['secondary_education_enrollment_pct'].str.replace('%','').astype(float)

# review changes
type_change('secondary_education_enrollment_pct')

# drop and rename
world_df=world_df.drop(columns='secondary_education_enrollment_pct').rename(columns={
    'secondary_education_enrollment_pct_edited':'secondary_education_enrollment_pct'})

# review
name_change('secondary_education_enrollment_pct')

***secondary_education_enrollment_pct** completed*

**population**

object - int

# display format
format('population')

# remove commas and convert to int
world_df['population_edited']=world_df['population'].str.replace(',','').astype(float)

# review chnage
type_change('population')

# drop, rename
world_df=world_df.drop(columns='population').rename(columns={'population_edited':'population'})

# review
name_change('population')

***population** completed*

**tax_revenue_pct**

object - float

# display format
format('tax_revenue_pct')

# remove % and convert
world_df['tax_revenue_pct_edited']=world_df['tax_revenue_pct'].str.replace('%','').astype(float)

# review
type_change('tax_revenue_pct')

# drop and rename
world_df=world_df.drop(columns='tax_revenue_pct').rename(columns={'tax_revenue_pct_edited':'tax_revenue_pct'})

# review
name_change('tax_revenue_pct')

***tax_revenue_pct** completed*

**unemployment_rt**

object - float

# display format
format('unemployment_rt')

# remove %
world_df['unemployment_rt_edited']=world_df['unemployment_rt'].str.replace('%','').astype(float)

# review
type_change('unemployment_rt')

# drop and rename
world_df=world_df.drop(columns='unemployment_rt').rename(columns={'unemployment_rt_edited':'unemployment_rt'})

# review
name_change('unemployment_rt')

***unemployment_rt** completed*

# last data type check on whole data frame
print(f'Data Types (after cleaning)\n-----------------------------\n{world_df.dtypes}')

---
### Save Data
* **world_data_cleaned.csv**

# display head one last time to ensure correct df
world_df.head()

# save to csv
world_df.to_csv('03_world_data_clean.csv',sep=',',encoding='utf-8',index=False,header=True)
