==> code/notebooks/03_BAIS3250_project_mergingData.ipynb <==
# BAIS:3250 - Final Project
### Merge Datasets

**Author(s):** Natalie Brown, Max Kaiser

**Date Modified:** 11-15-2024 (*date created:* 11-15-2024)


**Description:** Merge the country data from Kaggle with the country data from the API

---

### Import Libaries
* **pandas:** for data frames and data cleaning functions

import pandas as pd

---
### Load Data
* **world_data_clean.csv** - Kaggle data
* **api_country_data_clean.csv** - API data

# load data frames
kaggle_df=pd.read_csv('03_world_data_clean.csv',sep=',',encoding='utf-8')
api_df=pd.read_csv('01_api_country_data_clean.csv',sep=(','),encoding='utf-8')

# display kaggle data
kaggle_df.head()

# display api data
api_df.head()

---
#### Prepare Data Frames for Merging
- [x] Get unique country names for both and ensure the formatting matches

# create list of unique countries in kaggle df
kaggle_country_list=kaggle_df['country'].unique()

# display
print(kaggle_country_list)

# create list of unique countries in api df
api_country_list=api_df['country'].unique()

# display
print(api_country_list)

**there is an encoding error for one of the countries in the kaggle_df**
* **error:** Sï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½
- [x] compare the error row in kaggle_df with countries in the api_df
- [x] Could not find, but I believe it is Sao Tome and Principe because it would come after San Marino in alphabteical order, small discrepancy in population due to data collection time, but good enough so replace with Sao Tome and Principe
- [x] Create a map to replace names in the kaggle df to the format in api_df

# display api
api_df[api_df['country']=='Sao Tome and Principe']

# display error
kaggle_df[kaggle_df['country']=='Sï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½']

# loop through countries and find those that don't have match 
unique_in_api=[country for country in api_country_list if country not in kaggle_country_list]

# Find countries in kaggle_country_list not in api_country_list
unique_in_kaggle=[country for country in kaggle_country_list if country not in api_country_list]

# Print the countries that do not have a match
print("Countries in API list not in Kaggle list:")
for country in unique_in_api:
    print(country)

print("\nCountries in Kaggle list not in API list:")
for country in unique_in_kaggle:
    print(country)


# map to apply to kaggle, note we have to deal with congo rows, find way to merge the rows
country_map_kaggle={
    'The Bahamas':'Bahamas',
    'Republic of the Congo':'Congo',
    'Democratic Republic of the Congo':'Congo',
    'The Gambia':'Gambia',
    'Sï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½':'Sao Tome and Principe',
    'Republic of Ireland':'Ireland',
    'Federated States of Micronesia':'Micronesia',
}
    
# map to apply to api
country_map_api={
    'Cote D\'Ivoire':'Ivory Coast',
    'Antigua And Barbuda':'Antigua and Barbuda',
    'Trinidad And Tobago':'Trinidad and Tobago',
    'Saint Vincent And The Grenadines':'Saint Vincent and the Grenadines',
    'Swaziland':'Eswatini'
}


# apply map to kaggle
kaggle_df['country']=kaggle_df['country'].replace(country_map_kaggle)

# review
kaggle_df['country']

# apply map to api
api_df['country']=api_df['country'].replace(country_map_api)

# review
api_df['country']

**run loops to check country names again for mismatches**

# create lists for country edited  
api_country_list_edited=api_df['country'].unique()
kaggle_country_list_edited=kaggle_df['country'].unique()

# loop through countries and find those that don't have match 
unique_in_api_edited=[country for country in api_country_list_edited if country not in kaggle_country_list_edited]

# Find countries in kaggle_country_list not in api_country_list
unique_in_kaggle_edited=[country for country in kaggle_country_list_edited if country not in api_country_list_edited]

# Print the countries that do not have a match
print("Countries in API list not in Kaggle list:")
for country in unique_in_api_edited:
    print(country)

print("\nCountries in Kaggle list not in API list:")
for country in unique_in_kaggle_edited:
    print(country)


**review number of unique countries in kaggle_df to api_df**

- The difference in unique countries matches the number of countries that don't match
- These will not have matches in the final df, and therefor will be dropped

# get length
num_unique_countries_kaggle=len(kaggle_df['country'].unique())
num_unique_countries_api=len(api_df['country'].unique())

# display
print(f'Unique Countries in Kaggle: {num_unique_countries_kaggle}\nUnique Countries in API: {num_unique_countries_api}')

### Merge data frames with left join, will impute missing values for api data 

- [x] review join

# merge
merged_df=pd.merge(kaggle_df,api_df,on='country',how='left')

# display
merged_df.head()

# review shape
print(f'Merged df rows: {merged_df.shape[0]}\nMerged df columns: {merged_df.shape[1]}')

### Save data frame to csv file

---
### Save Data
* **new csv file name**

# save
merged_df.to_csv('04_country_merged_raw.csv',sep=',',encoding='utf-8',index=False)
