==> code/notebooks/05_BAIS3250_project_imputingCleanedData.ipynb <==
# BAIS:3250 - Final Project
### Data Transformation

**Author(s):** Natalie Brown, Max Kaiser

**Date Modified:** 11-21-2024 (*date created:* 11-21-2024)


**Description:** Transforming columns and dropping / imputing nulls for final dataset

---

### Import Libaries
* **pandas:** for data frames and data cleaning functions

import pandas as pd

---
### Load Data
* **country_merged_clean.csv**

# load data
country_df=pd.read_csv('05_country_merged_clean.csv',sep=',',encoding='utf-8')

# display
country_df.head()

# List of countries to drop 
countries_to_drop = ['North Macedonia', 'Palestinian National Authority', 'Vatican City'] 

# Drop rows with these countries 
country_df=country_df[~country_df['country'].isin(countries_to_drop)]

---
### Null Display
* determine how many nulls are in each column, should they be imputed or dropped?
    * we will look up data for nulls on WHO website and create a map to impute (world health organization) and other various reputable sources
* nulls columns:
  - [x] lat
  - [x] long
  - [x] infant mortality_rt
  - [x] birth_rt
  - [x] ~ag_land_pct~
      * we can drop this because we have data to compute a new column ourselves
  - [x] physicians per thousand
  - [x] fertility rt
  - [x] primary education enrollment_pct
  - [x] life expectancy
  - [x] secondary_education_enrollment_pct
  - [x] ~currency_code~
      * drop 
  - [x] ~consuer_price_index~
      * drop
  - [x] unemployment_rt
  - [x] tax_revenue_pct
  - [x] continent
  - [x] ~official language~
      * drop
  - [x] covid deaths
  - [x] covid cases
        

# function for nulls, returns a dataframe with nulls
def df_nulls(df):
    null_counts=country_df.isna().sum()
    null_percentages=round((df.isna().sum()/df.shape[0])*100,2)

    # data frame for results
    nulls_df=pd.DataFrame({
        'Nulls':null_counts,
        'Nulls (%)':null_percentages
    })
    
    # display
    print("Country Data Nulls")
    display(nulls_df.sort_values(by='Nulls',ascending=False))

df_nulls(country_df)

# function for displaying all nulls in a column
def find_nulls(df,column):
    return df[df[column].isnull()][['country',column]]

def null_list(df,column):
    return df[df[column].isnull()]['country'].tolist()
    
# function to display column format
def col_format(df,col):
    format_list=df[col].unique()
    return format_list[0:5]

---
**latitude**

# find countrys with null
find_nulls(country_df,'lat')

# find format
col_format(country_df,'lat')

**latitude found on [GPS Coordinates](https://gps-coordinates.org/sao-tome-and-principe-latitude.php)**
- 0.203237

# impute
country_df.loc[country_df['country'] == 'Sao Tome and Principe', 'lat'] = 0.203237

# review
find_nulls(country_df,'lat')

---
**longitude**

# find countrys with null
find_nulls(country_df,'long')

**longitude found on [GPS Coordinates](https://gps-coordinates.org/sao-tome-and-principe-latitude.php)**
- 6.608357

# impute
country_df.loc[country_df['country'] == 'Sao Tome and Principe', 'long'] = 6.608357

# review
find_nulls(country_df,'long')

---
**infant mortality rate found on [CIA World Factbook](https://www.cia.gov/the-world-factbook/field/infant-mortality-rate/country-comparison/)**

# find nulls
find_nulls(country_df,'infant_mortality_rt')

# get current format
col_format(country_df,'infant_mortality_rt')

# define infant morality rates
impute_mortality_rts={
    'Eswatini':36.7,
    'Liechtenstein':3.9,
    'Nauru':7.6
}

# impute
for country,rt in impute_mortality_rts.items():
    country_df.loc[country_df['country']==country,'infant_mortality_rt']=rt

# review
find_nulls(country_df,'infant_mortality_rt')

---
**birth rate found on [CIA World Factbook](https://www.cia.gov/the-world-factbook/field/birth-rate/country-comparison/)**

# find nulls
find_nulls(country_df,'birth_rt')

# get format
col_format(country_df,'birth_rt')

# define rates
impute_birth_rts={
    'Eswatini':22.3,
    'Nauru':20.2,
    'Tuvalu':22,
}

# impute
for country,rt in impute_birth_rts.items():
    country_df.loc[country_df['country']==country,'birth_rt']=rt

# review
find_nulls(country_df,'birth_rt')

---
**fertility rate found on [CIA World Factbook](https://www.cia.gov/the-world-factbook/field/total-fertility-rate/country-comparison/)**

# find nulls
find_nulls(country_df,'fertility_rt')

# get format
col_format(country_df,'fertility_rt')

# define rates
impute_fertility_rts={
    'Eswatini':2.37,
    'Monaco':1.54,
    'Nauru':2.55,
    'Tuvalu':2.78
}

# impute
for country,rt in impute_fertility_rts.items():
    country_df.loc[country_df['country']==country,'fertility_rt']=rt

# review
find_nulls(country_df,'fertility_rt')

---
**unemployment rate found on [CIA World Factbook](https://www.cia.gov/the-world-factbook/field/unemployment-rate/country-comparison/)**

# find nulls
find_nulls(country_df,'unemployment_rt')

# get format
col_format(country_df,'unemployment_rt')

# dictionary
impute_ue_rts={
    'Andorra': 3.7,
    'Antigua and Barbuda': 11,
    'Dominica': 11,
    'Eswatini': 37.64,
    'Grenada': 24,
    'Kiribati': 8.6,
    'Liechtenstein': 2.4,
    'Marshall Islands': 9.8,
    'Micronesia': 8.9, 
    'Monaco': 6.3,
    'Nauru': 5.1,
    'Palau': 1.7, 
    'San Marino': 8.1, 
    'Seychelles': 3, 
    'Tuvalu': 11,
    'Saint Kitts and Nevis':5.09
}

# impute
for country,rt in impute_ue_rts.items():
    country_df.loc[country_df['country']==country,'unemployment_rt']=rt

# review
find_nulls(country_df,'unemployment_rt')

---
**tax revenue percentage found on [CIA World Factbook](https://www.cia.gov/the-world-factbook/field/taxes-and-other-revenues/)**

# find nulls
find_nulls(country_df,'tax_revenue_pct')

# null list
null_list(country_df,'tax_revenue_pct')

# get format
col_format(country_df,'tax_revenue_pct')

# define tax revenues
impute_tax_revenue_percentages={
    'Andorra': 69,
    'Brunei': 18.5,
    'Chad': 13.08,
    'Comoros': 14.22,
    'Cuba': 58.1,
    'Djibouti': 18.95,
    'Ecuador': 13.0,
    'Eritrea': 34.9,
    'Guyana': 15.37,
    'Haiti': 6.24,
    'Libya': 51.6,
    'Liechtenstein': 52.3,
    'Mauritania': 27.4,
    'Monaco': 14.4,
    'Montenegro': 37.2,
    'Nauru': 44.3,
    'Panama': 7.5,
    'South Sudan': 8.5,
    'Turkmenistan': 14.9,
    'Tuvalu': 18.5,
    'Venezuela': 5.96,
    'Yemen': 9.58,
    'North Korea':0 # abolished taxing
    }

# impute
for country,pct in impute_tax_revenue_percentages.items():
    country_df.loc[country_df['country']==country,'tax_revenue_pct']=pct

# review
find_nulls(country_df,'tax_revenue_pct')


---
**life expectancy found on [CIA World Factbook](https://www.cia.gov/the-world-factbook/field/life-expectancy-at-birth/country-comparison/)**

# find nulls
find_nulls(country_df,'life_expectancy')

# list of countries
null_list(country_df,'life_expectancy')

# get format
col_format(country_df,'life_expectancy')

# define list to impute
impute_life_expectancies={
    'Andorra':83.8,
    'Eswatini':60.7,
    'Monaco':89.8,
    'Nauru':68.6,
    'Tuvalu':69
}

# impute
for country,expectancy in impute_life_expectancies.items():
    country_df.loc[country_df['country']==country,'life_expectancy']=expectancy

# review
find_nulls(country_df,'life_expectancy')

---
**physicians per thousand found on [CIA World Factbook](country_df,'physicians_per_thousand')**

# find nulls
find_nulls(country_df,'physicians_per_thousand')

# list of nulls
null_list(country_df,'physicians_per_thousand')

# get format
col_format(country_df,'physicians_per_thousand')

# define dictionary to impute
impute_phys_per_thousand={
    'Eswatini':0.14,
    'Liechtenstein':2.5,
    'Nauru':1.35,
    'South Sudan':0.04
}

# impute
for country,stat in impute_phys_per_thousand.items():
    country_df.loc[country_df['country']==country,'physicians_per_thousand']=stat

# review
find_nulls(country_df,'physicians_per_thousand')


---
**primary education enrollment percentage found on [Our World in Data](https://ourworldindata.org/grapher/total-net-enrollment-rate-in-primary-education?tab=table&time=2021..latest)**

# find nulls
find_nulls(country_df,'primary_education_enrollment_pct')

# list
null_list(country_df,'primary_education_enrollment_pct')

# format
col_format(country_df,'primary_education_enrollment_pct')

# define dictionary to impute
prim_enrollment_pct={
    'Bosnia and Herzegovina':84.4,
    'Eswatini':89.5,
    'Monaco':97.9,
    'Nauru':96.3
}

# impute
for country,pct in prim_enrollment_pct.items():
    country_df.loc[country_df['country']==country,'primary_education_enrollment_pct']=pct

# review
find_nulls(country_df,'primary_education_enrollment_pct')

---
**secondary education enrollment percentage found on [Our World in Data](https://ourworldindata.org/grapher/total-net-enrollment-rate-in-primary-education?tab=table&time=2021..latest)**

# find nulls
find_nulls(country_df,'secondary_education_enrollment_pct')

# get list
null_list(country_df,'secondary_education_enrollment_pct')

# define dictionary to impute
impute_secd_enrollment_pct={
    'Andorra':97.5,
    'Bolivia':91.5,
    'Eswatini':86,
    'Kiribati':81.6,
    'Monaco':154.2,
    'Nauru':86.4,
    'Solomon Islands':48.3,
    'South Sudan':11.2,
    'Tuvalu':91.3
}

# impute
for country,pct in impute_secd_enrollment_pct.items():
    country_df.loc[country_df['country']==country,'secondary_education_enrollment_pct']=pct

# review
find_nulls(country_df,'secondary_education_enrollment_pct')

---
**continent**

# find nulls
find_nulls(country_df,'continent')

# get list
null_list(country_df,'continent')

# define dictionary to impute
impute_continent={
    'Cape Verde':'Africa',
    'Netherlands':'Europe',
    'North Korea':'Asia',
    'Saint Kitts and Nevis':'North America',
    'South Korea':'Asia',
    'East Timor':'Asia'
    
}

# impute
for country,continent in impute_continent.items():
    country_df.loc[country_df['country']==country,'continent']=continent

# review
find_nulls(country_df,'continent')

---
**covid deaths**

# find nulls
find_nulls(country_df,'covid_deaths')

# find nulls
null_list(country_df,'covid_deaths')

# define dictionary to impute
impute_deaths={
    'Cape Verde':417,
     'Netherlands':22986,
     'North Korea':74,
     'Saint Kitts and Nevis':0,
     'South Korea':20000,
     'East Timor':0

    
}

# impute
for country,death in impute_deaths.items():
    country_df.loc[country_df['country']==country,'covid_deaths']=death

# review
find_nulls(country_df,'covid_deaths')

---
**covid cases**


# find nulls
find_nulls(country_df,'covid_cases')

# find nulls
null_list(country_df,'covid_cases')

# define dictionary to impute
impute_cases={
    'Cape Verde':64474,
    'Netherlands':8300000,
    'North Korea':4772813,
    'Saint Kitts and Nevis':0,
    'South Korea':14500000,
    'East Timor':0

    
}

# impute
for country,case in impute_cases.items():
    country_df.loc[country_df['country']==country,'covid_cases']=case

# review
find_nulls(country_df,'covid_cases')

---
**drop columns**

country_df=country_df.drop(columns=['ag_land_pct','official_language','consumer_price_index','currency_code','total_area_km2','land_area_km2'])

---
**check nulls before saving**

df_nulls(country_df)

---
### Save Data
* **new csv file name**

# save csv file
country_df.to_csv('06_country_merged_imputed.csv',sep=',',encoding='utf-8',index=False)
