==> code/notebooks/06_BAIS3250_project_transformingCleanedData.ipynb <==
# BAIS:3250 - Final Project
### Transforming Columns for Analysis

**Author(s):** Natalie Brown, Max Kaiser

**Date Modified:** 12-13-2024 (*date created:* 12-02-2024)


**Description:** Transforming various columns for a seamless analysis across countries

---

### Import Libaries
* **pandas:** for data frames and data cleaning functions

import pandas as pd
import numpy as np

---
### Load Data
* **06_country_merged_imputed.csv**

# load cleaned and imputed csv file
country_df=pd.read_csv('06_country_merged_imputed.csv',sep=',',encoding='utf-8')

# view
country_df.head()

---
### Address Data Types Before Transformation

# data types
print(f'Country df Data Types:\n{country_df.dtypes}')

### Transform Columns
- [x] make df copy for transformations
- [x] make covid death rate column (covid deaths / covid cases)
    - [x] drop columns after     
- [x] rename population

---

**covid death rate**

# make copy
country_transformed_df=country_df

# make columns accounting for those with zero 
country_transformed_df['covid_death_rt']=np.where(
    (country_transformed_df['covid_deaths'] == 0) | (country_transformed_df['covid_cases'] == 0),
    0,
    round(country_transformed_df['covid_deaths']/country_transformed_df['covid_cases'],4)*100
)

# review
country_transformed_df[['country','covid_death_rt']]

# drop
country_transformed_df=country_transformed_df.drop(columns=['covid_cases','covid_deaths'])

---
**rename population**

# rename
country_transformed_df=country_transformed_df.rename(columns={'average_population':'population'})

---
### Re-Order Columns

# get current order
country_transformed_df.columns

# reorder
country_transformed_df=country_transformed_df[['country','population','lat','long','continent',
                                               'gross_domestic_product_usd(b)','tax_revenue_pct','unemployment_rt',
                                               'primary_education_enrollment_pct','secondary_education_enrollment_pct',
                                               'life_expectancy','birth_rt','infant_mortality_rt','physicians_per_thousand','covid_death_rt']]

# review
country_transformed_df.head()

---
### Save Data
* **07_country_final.csv**

# save
country_transformed_df.to_csv('07_country_final.csv',sep=',',encoding='utf-8',index=False)
